+++ src/Symfony/Component/Translation/Dumper/IniFileDumper.php
@@ -33,7 +33,7 @@ class IniFileDumper extends FileDumper
     /**
      * {@inheritdoc}
      */
-    protected function formatCatalogue(MessageCatalogue $messages, $domain, array $options = array())
+    public function formatCatalogue(MessageCatalogue $messages, $domain, array $options = array())
     {
         $output = '';