+++ src/Symfony/Component/Translation/Dumper/QtFileDumper.php
@@ -33,7 +33,7 @@ class QtFileDumper extends FileDumper
     /**
      * {@inheritdoc}
      */
-    protected function formatCatalogue(MessageCatalogue $messages, $domain, array $options = array())
+    public function formatCatalogue(MessageCatalogue $messages, $domain, array $options = array())
     {
         $dom = new \DOMDocument('1.0', 'utf-8');
         $dom->formatOutput = true;