+++ src/Symfony/Component/Translation/Tests/Dumper/JsonFileDumperTest.php
@@ -16,19 +16,7 @@ use Symfony\Component\Translation\Dumper\JsonFileDumper;
 
 class JsonFileDumperTest extends \PHPUnit_Framework_TestCase
 {
-    private $tempDir;
-
-    protected function setUp()
-    {
-        $this->tempDir = sys_get_temp_dir();
-    }
-
-    protected function tearDown()
-    {
-        unlink($this->tempDir.'/messages.en.json');
-    }
-
-    public function testDump()
+    public function testFormatCatalogue()
     {
         if (PHP_VERSION_ID < 50400) {
             $this->markTestIncomplete('PHP below 5.4 doesn\'t support JSON pretty printing');
@@ -38,9 +26,8 @@ class JsonFileDumperTest extends \PHPUnit_Framework_TestCase
         $catalogue->add(array('foo' => 'bar'));
 
         $dumper = new JsonFileDumper();
-        $dumper->dump($catalogue, array('path' => $this->tempDir));
 
-        $this->assertEquals(file_get_contents(__DIR__.'/../fixtures/resources.json'), file_get_contents($this->tempDir.'/messages.en.json'));
+        $this->assertStringEqualsFile(__DIR__.'/../fixtures/resources.json', $dumper->formatCatalogue($catalogue, 'messages'));
     }
 
     public function testDumpWithCustomEncoding()
@@ -49,8 +36,7 @@ class JsonFileDumperTest extends \PHPUnit_Framework_TestCase
         $catalogue->add(array('foo' => '"bar"'));
 
         $dumper = new JsonFileDumper();
-        $dumper->dump($catalogue, array('path' => $this->tempDir, 'json_encoding' => JSON_HEX_QUOT));
 
-        $this->assertEquals(file_get_contents(__DIR__.'/../fixtures/resources.dump.json'), file_get_contents($this->tempDir.'/messages.en.json'));
+        $this->assertStringEqualsFile(__DIR__.'/../fixtures/resources.dump.json', $dumper->formatCatalogue($catalogue, 'messages', array('json_encoding' => JSON_HEX_QUOT)));
     }
 }